Make tls dependencies optional
