+++ rumqttd-ng/Cargo.toml
@@ -16,15 +16,14 @@ slab = "0.4.3"
 log = "0.4.14"
 thiserror = "1.0.24"
 tokio-util = { version = "0.7", features = ["codec"] }
-tokio-rustls =  "0.23.0"
+tokio-rustls =  { version = "0.23.0", optional = true }
 tokio-native-tls = { version = "0.3", optional = true }
-rustls-pemfile = "0.3.0"
-webpki = "0.22.0"
+rustls-pemfile = { version = "0.3.0", optional = true }
 tokio-tungstenite = "0.15.0"
 ws_stream_tungstenite = "0.8"
 websocket-codec = "0.5.1"
 rouille = "3.1.1"
-x509-parser = "0.9.2"
+x509-parser = {version= "0.9.2", optional = true}
 futures-util = "0.3.16"
 parking_lot = "0.11.2"
 reqwest = { version = "0.11", features = ["json"] }
@@ -34,7 +33,9 @@ simplelog = "0.12.0"
 structopt = "0.3.26"
 
 [features]
-native-tls = ["tokio-native-tls"]
+default = ["use-rustls"]
+use-rustls = ["tokio-rustls", "rustls-pemfile", "x509-parser"]
+use-native-tls = ["tokio-native-tls", "x509-parser"]
 
 [dev-dependencies]
 pretty_env_logger = "0.4.0"
